Replaced LazyLoader with functional loader module
